cleanup upcoming merge (previous commit_)
this merge added back in a function GetOnOffBike, so it needed to be added back to bike.h as well. I'm not sure if it gets called or does anything since I added outfits, but it doesn't seem to get in the way of anything. BS_TARGET needed to be changed to BS_SCRIPTING to stay consistent between normal BattleScirpt_FaintBattler and BattleScript_Faint(Groudon/Kyogre)