added mr briney pokenav stuff
mr briney is now registered in pokenav after first boat trip to slateport, he can be called from route 104 or 109 when he is at the opposite and he will come to where you are to pick you up. there are messages for every relevant combination of player/briney locations including after beating gym 5. modified some of the soft lock prevention on briney, but I think I compensated sufficiently.  all scripts and code work as intended but might have extra unnecessary stuff because I was tryialing and erroring to get it to work.